+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 *(Please use the [release template](docs/dev/release/release-notes-template.md) as the template for this document)*
 
+# MongoDB Enterprise Kubernetes Operator 1.16.5
+
+* Improved security by introducing `readOnlyRootFilesystem` property to all deployed containers. This change also introduces a few additional volumes and volume mounts. 
+
 # MongoDB Enterprise Kubernetes Operator 1.16.4
 
 ## Security fixes